Refactor: Move ChatMessages classes into new chatMessages package
